+++ lib/eve_online.rb
@@ -48,6 +48,7 @@ require 'eve_online/esi/market_history'
 
 require 'eve_online/esi/corporation'
 require 'eve_online/esi/corporation_industry_jobs'
+require 'eve_online/esi/corporation_divisions'
 
 require 'eve_online/esi/alliance'
 require 'eve_online/esi/alliance_icon'
@@ -79,6 +80,7 @@ require 'eve_online/esi/models/server_status'
 require 'eve_online/esi/models/market_history'
 
 require 'eve_online/esi/models/corporation'
+require 'eve_online/esi/models/division'
 
 require 'eve_online/esi/models/alliance'
 require 'eve_online/esi/models/alliance_icon'