+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Changelog
 
 **master**
+* Add `EveOnline::ESI::Models::Division`
+* Add `EveOnline::ESI::CorporationDivisions`
 * Add `EveOnline::ESI::Models::MarketHistory`
 * Add `EveOnline::ESI::MarketHistory`
+
 
 **v0.13.0**